Refactored/renamed delegates reduce code duplication; added mistune markdown library
