+++ delegates/CurationItem.ui
